Better error on compiling null regex
It is possible for users to construct a regex that matches the null set, i.e.
no strings, not even the empty string. An example is `re"A" & re"B"`.
Automa could support these, but don't currently, as there is no point to.
However, attempting to compile such a regex currently throws an obscure internal
error.

This PR improves that error.

See issue #104