Also check ambiguities for final and all actions
An oversight in the ambiguity check meant that actions placed on non-epsilon
edges were accidentally not included in the paths for validation.
MWE: `compile(onfinal!(re"a", :a) | onfinal!(re"a", :b))`

This breaks tokenizers, so we manually skip ambiguity check in tokenizers.
In the case of conflicting actions in tokenizers, this will cause the longest
matching token to be emitted.