Do not assert mark has set with special macros
Special macros like `@abspos` previously inserted an assert that the buffer's
mark had been set. Unfortunately, this error branch made the macro unsuitable
in performance senstitive locations.
Now simply skip the check. This will cause undefined behaviour if the mark is
not set (in that the macro will evaluate to a wrong integer, not crazy
nasal demon UB), but that should be OK.