Backport from v1: Remove SIMD capability
I've come to the conclusion that Julia does not make it possible to robustly
check what CPU instructions the user has available.
The current options are all undocumented, complex and brittle, and not suitable
for code that cannot be accepted to break at any time

Whenever a robust way of checking for CPU instructions are available, the change
is easy to revert.